Add files via upload
The program opens up a calculator built by using a GridPane to format where the user is asked to input two numbers in Textfields along with buttons used to perform the 4 basic operations: add, subtract, multiply, and divide and also pressing the clear button to clear both the input fields. Once the user enters the two numbers and presses on each button, it will produce the result. If the user enters anything other than a floating point number, then the result is Invalid. For division, if a user divides anything by zero, the result will display, invalid/Infinity as the answer is undefined.